:recycle: move redirect.js in router
