+++ src/store/reducers/auth/protect.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { inject } from '@k-ramel/react'
-import { redirectWithNextUrl } from 'helpers/redirect'
+import { redirectWithNextUrl } from 'store/drivers/router/redirect'
 
 import LoadingIndicator from 'components/loader/loading'
